+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Fixed size issue with iPhone apps running @2x on an iPad (see `row_type/base.rb#field_buffer`)
 
+- Fixed bug in `PickerRow` where KVOing `:value` changes woulnd't reflect in the picker UI.
+
 ## 1.1.5 - November 9, 2012
 
 ### Features